cast integer to size_t for comparison

+++ 2-selection_sort.c
@@ -9,7 +9,7 @@ void selection_sort(int *array, size_t size)
 {
 	size_t i = 0, j = 0, min = 0, holder = 0;
 
-	if (size < 2)
+	if (size < (size_t)2)
 		return;
 
 	for (i = 0; i < size; i++)
